third_party/blink/web_tests/platform/generic/fast/events/page-visibility-null-view-expected.txt

This test checks that Page Visibility state values are correct when a document has no defaultView.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS subdocument.defaultView == null is true
PASS subdocument.visibilityState is "hidden"
PASS subdocument.hidden is true
PASS successfullyParsed is true

TEST COMPLETE
